deps: remove extraneous 'path' dependency
The 'path' module from core should be used, not a polyfil. Additionally,
by having this as a top level dependency, any of this package's
dependencies will get the module version isntead of the one from core.
And with npm3's dependency flattening, this could extend all the way to
the top-level of someone's app and end up injecting an unwanted
replacement for 'path' for the entire app.