fix: add tests for custom column name for id
These are test cases for a fix to loopback-ibmdb.

See issue https://github.com/strongloop/loopback-ibmdb/pull/80
for details.

If a 'columnName' value option is specified for the id
property of a model, it should be used in the generated
SQL. Otherwise, the model name for the property id
should be used in the generated SQL.